feat(state): implement `changed_callback` for `Exe` (CAN BE REVERTED)
WARNING: This is a dangerous commit.

The `Exe` structure now holds *raw mutable pointers* to `MarkovState`.
This means, we have to be extremely careful when dealing with anything
related to `MarkovState`.

I honestly have no idea how dangerous this commit is. But I hope that
the `impl Drop` will protect the program from unwanted `SEGFAULT`.